minor CR comments addressed
